Updated gencov.rb, gendoc.rb, added runlint.rb, run all three in per-commit hook.

+++ gencov.rb
@@ -1,12 +1,14 @@
 #!/usr/local/opt/ruby/bin/ruby
 
+# Assumes working directory is project root
+
 # Run the unit tests
 
 print "Running unit tests...\n"
 buildOut = `xcodebuild test -scheme Strux-Package -enableCodeCoverage YES -derivedDataPath .build/derivedData`
 buildMatch = buildOut.match(/Test session results, code coverage, and logs:\s+(?<file>\S+)\s+.*TEST SUCCEEDED/)
 if !!!buildMatch
-    print "Test failed!\n"
+    print "test exit - unit tests failed 😞\n"
     exit 1
 end
 
@@ -15,9 +17,6 @@ text_file = "/tmp/codecov.html"
 badge_file = "/tmp/codecov.svg"
 
 # Generate the coverage report
-
-#system "xcrun xccov view --report #{result_file} > #{text_file}"
-#covOut = File.read(text_file)
 
 pre_body = <<-PRE_BODY
 <!DOCTYPE html>
@@ -51,7 +50,7 @@ pct = ""
 targets.each do | target |
     targetMatch = target.match(/^(?<name>\S+)\s+(?<coverage>\d+\.\d+)%/)
     if !!!targetMatch
-        print "Unable to parse coverage output!\n"
+        print "test exit - unable to parse coverage output 😞\n"
         exit 1
     end
     targetName = targetMatch[:name]
@@ -63,7 +62,7 @@ targets.each do | target |
     end
 end
 if pct == ""
-    print "Unable to parse coverage output!\n"
+    print "test exit - unable to parse coverage output 😞\n"
     exit 1
 end
 
@@ -83,10 +82,26 @@ elsif pctFloat >= 60
     color = "orange"
 end
 
-system "curl \"https://img.shields.io/badge/coverage-#{pct}%25-#{color}\" > #{badge_file}"
+if !system "curl \"https://img.shields.io/badge/coverage-#{pct}%25-#{color}\" > #{badge_file}"
+    print "test exit - couldn't create badge file 😞\n"
+    exit 1
+end
 system "gsutil rm gs://scipioapps/mystrux/codecov.html"
 system "gsutil rm gs://scipioapps/mystrux/codecov.svg"
-system "gsutil -h \"Cache-Control: no-cache\" cp #{text_file} gs://scipioapps/mystrux/codecov.html"
-system "gsutil -h \"Cache-Control: no-cache\" cp #{badge_file} gs://scipioapps/mystrux/codecov.svg"
-system "gsutil acl ch -u AllUsers:R gs://scipioapps/mystrux/codecov.html"
-system "gsutil acl ch -u AllUsers:R gs://scipioapps/mystrux/codecov.svg"
+if !system "gsutil -h \"Cache-Control: no-cache\" cp #{text_file} gs://scipioapps/mystrux/codecov.html"
+    print "test exit - couldn't write codecov.html 😞\n"
+    exit 1
+end
+if !system "gsutil -h \"Cache-Control: no-cache\" cp #{badge_file} gs://scipioapps/mystrux/codecov.svg"
+    print "test exit - couldn't write codecov.svg 😞\n"
+    exit 1
+end
+if !system "gsutil acl ch -u AllUsers:R gs://scipioapps/mystrux/codecov.html"
+    print "test exit - couldn't set acl for codecov.html 😞\n"
+    exit 1
+end
+if !system "gsutil acl ch -u AllUsers:R gs://scipioapps/mystrux/codecov.svg"
+    print "test exit - couldn't set acl for codecov.svg 😞\n"
+    exit 1
+end
+print "test success 🤠 YEEE HAW!\n"
